requirements.txt
streamlit==1.38.0
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
opencv-python==4.10.0.84
